requirements.txt
discord.py-self
python-dotenv
requests